+++ src/main/java/com/behrend/contestmanager/ContestManagerApplication.java
@@ -1,0 +1,13 @@
+package com.behrend.contestmanager;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ContestManagerApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ContestManagerApplication.class, args);
+    }
+    
+}
