--- src/main/java/com/behrend/sscontestmanager/SscontestmanagerApplication.java
@@ -1,13 +1,0 @@
-package com.behrend.sscontestmanager;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SscontestmanagerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SscontestmanagerApplication.class, args);
-	}
-
-}
