+++ src/test/java/com/behrend/contestmanager/ContestManagerApplicationTests.java
@@ -1,13 +1,13 @@
-package com.behrend.sscontestmanager;
+package com.behrend.contestmanager;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class SscontestmanagerApplicationTests {
+class ContestManagerApplicationTests {
 
-	@Test
-	void contextLoads() {
-	}
+    @Test
+    void contextLoads() {
+    }
 
 }
